% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-animation.R
\name{animate}
\alias{animate}
\title{Animate}
\usage{
animate(enter = NULL, update = NULL, exit = NULL)
}
\arguments{
\item{enter}{list(type=NULL,duration=NULL,delay=NULL,easing=NULL)}

\item{update}{list(type=NULL,duration=NULL,delay=NULL,easing=NULL)}

\item{exit}{list(type=NULL,duration=NULL,delay=NULL,easing=NULL)}
}
\description{
Animate in G2 is an important part of visualization
and can improve the expressiveness of visualization.
}
\details{
Animate can be declared at the level of mark.

Mark specifies animation properties through mark.animate,
there are three parts of animation that can be specified:
\itemize{
\item enter- New graphics
\item update- Updated graphics
\item exit- deleted graphics
Each part of the animation has the following properties:
\item type
\item duration
\item delay
\item easing
}

animate types:
\itemize{
\item fadeIn
\item fadeOut
\item growInX
\item growInY
\item morphing
\item pathIn
\item scaleInX
\item scaleInY
\item scaleOutX
\item scaleOutY
\item waveIn
\item zoomIn
\item zoomOut
}
}
\concept{animation}
