% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-component.R
\name{axis}
\alias{axis}
\title{Axis}
\usage{
axis(x = NULL, y = NULL)
}
\description{
Axis
}
\seealso{
Other component: 
\code{\link{legend}()},
\code{\link{style}()},
\code{\link{title}()}
}
\concept{component}
