% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-coordinate.R
\name{coordinate}
\alias{coordinate}
\title{Coordinate}
\usage{
coordinate(type = NULL, ...)
}
\arguments{
\item{type}{fisheye / parallel / polar / radial / theta / transpose / cartesian3D}

\item{transform}{array of transform.}
}
\description{
coordinate be configured in \code{g2()} or in \code{g2_} family functions.
}
\details{
Each view can only have one coordinate system. In addition to its own attributes,
the coordinate system also contains a series of Coordinate Transform.

The coordinate system of mark has bubbling properties and will be merged with
The coordinate system of view.
The coordinate system of the first mark has the highest priority.
}
