% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{encode}
\alias{encode}
\title{Encode}
\usage{
encode(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{encoding x}

\item{y}{encoding x}

\item{color}{encoding color}

\item{shape}{encoding shape}
}
\description{
encode can be configured either in mark or in view
}
\details{
For Bar charts, you will need to switch x and y encoding
}
