% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2}
\alias{g2}
\title{g2 plot}
\usage{
g2(data, chart, cfg = list(), width = NULL, height = NULL)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame,
data should only contains the necessary data for chart (reduce json/data size)}

\item{chart}{the chart type of G2Plot ('line','area','column','bar','pie','scatter','gauge','dual-axes','liquid','radar','word-cloud','funnel','bullet','histogram','venn','rose','tiny-line','tiny-area','tiny-column','sunburst','stock','ring-progress','progress','box','heatmap','waterfall','radial-bar','bidirectional-bar','sankey','chord','treemap','violin','circle-packing','mix','facet')}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.antgroup.com/en/api} for more details.}

\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}
}
\description{
all g2 plots can be confined in cfg param
}
\examples{
cfg = list(
  xField='drat',
  yField='wt',
  shape='circle',
  legend = list(position = 'top-left')
)
g2(mtcars, 'scatter', cfg = cfg)


}
\seealso{
\link{cfg}

Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()}
}
\concept{g2}
