% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_box}
\alias{g2_box}
\title{g2_box}
\usage{
g2_box(
  data,
  x,
  y,
  group = NULL,
  outliers = NULL,
  boxStyle = NULL,
  outliersStyle = NULL
)
}
\arguments{
\item{x}{xField}

\item{y}{yField}

\item{group}{Grouping field. It is used for grouping by default, and color is used as visual channel.}

\item{outliers}{Outlier field.}

\item{boxStyle}{list(fill="#f6f",...)}

\item{outliersStyle}{list(fill="#f6f",...)}
}
\description{
g2_box
}
\examples{
df = jsonlite::fromJSON('[
{ "x": "Oceania", "low": 1, "q1": 9, "median": 16, "q3": 22, "high": 24 },
{ "x": "East Europe", "low": 1, "q1": 5, "median": 8, "q3": 12, "high": 16 },
{ "x": "Australia", "low": 1, "q1": 8, "median": 12, "q3": 19, "high": 26 },
{ "x": "South America", "low": 2, "q1": 8, "median": 12, "q3": 21, "high": 28 },
{ "x": "North Africa", "low": 1, "q1": 8, "median": 14, "q3": 18, "high": 24 },
{ "x": "North America", "low": 3, "q1": 10, "median": 17, "q3": 28, "high": 30 },
{ "x": "West Europe", "low": 1, "q1": 7, "median": 10, "q3": 17, "high": 22 },
{ "x": "West Africa", "low": 1, "q1": 6, "median": 8, "q3": 13, "high": 16 }
]')
g2_box(df, 'x',c('low', 'q1', 'median', 'q3', 'high'))

# different data format (yField)
df = jsonlite::fromJSON('
[
{ "Species": "I. setosa", "type": "SepalLength", "value": 5.1, "bin": [4.3, 4.8, 5, 5.2, 5.8] },
{ "Species": "I. setosa", "type": "SepalWidth", "value": 3.5, "bin": [2.3, 3.2, 3.4, 3.7, 4.4] },
{ "Species": "I. setosa", "type": "PetalLength", "value": 1.4, "bin": [1, 1.4, 1.5, 1.6, 1.9] },
{ "Species": "I. setosa", "type": "PetalWidth", "value": 0.2, "bin": [0.1, 0.2, 0.2, 0.3, 0.6] },
{ "Species": "I. versicolor", "type": "SepalLength", "value": 7, "bin": [4.9, 5.6, 5.9, 6.3, 7] },
{ "Species": "I. versicolor", "type": "SepalWidth", "value": 3.2, "bin": [2, 2.5, 2.8, 3, 3.4] },
{ "Species": "I. versicolor", "type": "PetalLength", "value": 4.7, "bin": [3, 4, 4.35, 4.6, 5.1] },
{ "Species": "I. versicolor", "type": "PetalWidth", "value": 1.4, "bin": [1, 1.2, 1.3, 1.5, 1.8] },
{ "Species": "I. virginica", "type": "SepalLength", "value": 6.3, "bin": [4.9, 6.2, 6.5, 6.9, 7.9] },
{ "Species": "I. virginica", "type": "SepalWidth", "value": 3.3, "bin": [2.2, 2.8, 3, 3.2, 3.8] },
{ "Species": "I. virginica", "type": "PetalLength", "value": 6, "bin": [4.5, 5.1, 5.55, 5.9, 6.9] },
{ "Species": "I. virginica", "type": "PetalWidth", "value": 2.5, "bin": [1.4, 1.8, 2, 2.3, 2.5] }
]')
g2_box(df, 'type','bin',group = 'Species')

}
