% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-mark.R
\name{g2_cell}
\alias{g2_cell}
\title{cell}
\usage{
g2_cell(encode = NULL, ...)
}
\arguments{
\item{encode}{x / y / z / color / shape / size ...}
}
\description{
cell
}
\details{
Shape:
\itemize{
\item cell
\item hollow
}

Style:
\itemize{
\item radius
\item radiusTopLeft
\item radiusTopRight
\item radiusBottomRight
\item radiusBottomLeft
\item inset
\item insetLeft
\item insetRight
\item insetBottom
\item insetTop
\item fill
\item fillOpacity
\item stroke
\item strokeOpacity
\item lineWidth
\item lineDash
\item opacity
\item shadowColor
\item shadowBlur
\item shadowOffsetX
\item shadowOffsetY
\item cursor
}
}
\seealso{
Other mark: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_box}()},
\code{\link{g2_boxplot}()},
\code{\link{g2_chord}()},
\code{\link{g2_col}()},
\code{\link{g2_density}()},
\code{\link{g2_gauge}()},
\code{\link{g2_heatmap}()},
\code{\link{g2_hist}()},
\code{\link{g2_image}()},
\code{\link{g2_interval}()},
\code{\link{g2_line}()},
\code{\link{g2_lineX}()},
\code{\link{g2_lineY}()},
\code{\link{g2_link}()},
\code{\link{g2_liquid}()},
\code{\link{g2_pie}()},
\code{\link{g2_point}()},
\code{\link{g2_polygon}()},
\code{\link{g2_range}()},
\code{\link{g2_range_x}()},
\code{\link{g2_range_y}()},
\code{\link{g2_rect}()},
\code{\link{g2_shape}()},
\code{\link{g2_text}()},
\code{\link{g2_vector}()},
\code{\link{g2_wordCloud}()}
}
\concept{mark}
