% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_chord}
\alias{g2_chord}
\title{g2_chord}
\usage{
g2_chord(data, sourceField, targetField, weightField)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{sourceField}{source field}

\item{targetField}{target field}

\item{weightField}{weight field}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt using function with \code{g2_} prefix and \code{g2_opt()}.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
\examples{
data = jsonlite::fromJSON('[
{ "source": "北京", "target": "天津", "value": 30 },
{ "source": "北京", "target": "上海", "value": 80 },
{ "source": "北京", "target": "河北", "value": 46 },
{ "source": "北京", "target": "辽宁", "value": 49 },
{ "source": "北京", "target": "黑龙江", "value": 69 },
{ "source": "北京", "target": "吉林", "value": 19 },
{ "source": "天津", "target": "河北", "value": 62 },
{ "source": "天津", "target": "辽宁", "value": 82 },
{ "source": "天津", "target": "上海", "value": 16 },
{ "source": "上海", "target": "黑龙江", "value": 16 },
{ "source": "河北", "target": "黑龙江", "value": 76 },
{ "source": "河北", "target": "内蒙古", "value": 24 },
{ "source": "内蒙古", "target": "北京", "value": 32 }
]')
g2_chord(data, 'source','target', 'value') |> g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_bidirectional_bar}()},
\code{\link{g2_box}()},
\code{\link{g2_bullet}()},
\code{\link{g2_circle_packing}()},
\code{\link{g2_column}()},
\code{\link{g2_dual_axes}()},
\code{\link{g2_funnel}()},
\code{\link{g2_gauge}()},
\code{\link{g2_heatmap}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_liquid}()},
\code{\link{g2_pie}()},
\code{\link{g2_progress}()},
\code{\link{g2_radar}()},
\code{\link{g2_radial_bar}()},
\code{\link{g2_ring_progress}()},
\code{\link{g2_rose}()},
\code{\link{g2_sankey}()},
\code{\link{g2_scatter}()},
\code{\link{g2_stock}()},
\code{\link{g2_sunburst}()},
\code{\link{g2_tiny_area}()},
\code{\link{g2_tiny_column}()},
\code{\link{g2_treemap}()},
\code{\link{g2_venn}()},
\code{\link{g2_violin}()},
\code{\link{g2_waterfall}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2}()}
}
\concept{g2}
