% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_column}
\alias{g2_column}
\title{g2_column}
\usage{
g2_column(
  data,
  xField,
  yField,
  seriesField = NULL,
  groupField = NULL,
  colorField = NULL,
  isGroup = NULL,
  isStack = NULL,
  isRange = NULL,
  isPercent = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{xField}{the name of the data field corresponding to the graph in the x direction. you should config the \code{opt} param with \code{xField} in \code{g2()} function, and \code{x} (short for \code{xField}) is param name for functions with \code{g2_} prefix.}

\item{yField}{the name of the data field corresponding to the graph in the y direction.}

\item{isGroup}{to group charts}

\item{isStack}{to stack charts}

\item{isRange}{to range charts}

\item{isPercent}{to percent charts}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{
df = jsonlite::fromJSON('https://gw.alipayobjects.com/os/antfincdn/8elHX\%26irfq/stack-column-data.json',simplifyVector = TRUE)
g2_column(df,'year','value','type',isStack=TRUE)

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_funnel}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2_scatter}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2plot}()}
}
\concept{g2}
