% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_facet_circle}
\alias{g2_facet_circle}
\title{facetCircle}
\usage{
g2_facet_circle(data, encode, children, title, scale)
}
\description{
facetCircle
}
\seealso{
Other composition: 
\code{\link{g2_facet_rect}()},
\code{\link{g2_repeat_matrix}()},
\code{\link{g2_space_flex}()},
\code{\link{g2_space_layer}()},
\code{\link{g2_timing_keyframe}()}
}
\concept{composition}
