% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_facet_rect}
\alias{g2_facet_rect}
\title{facetRect}
\usage{
g2_facet_rect(data, encode, children, title, scale)
}
\description{
facetRect
}
\seealso{
Other composition: 
\code{\link{g2_facet_circle}()},
\code{\link{g2_repeat_matrix}()},
\code{\link{g2_space_flex}()},
\code{\link{g2_space_layer}()},
\code{\link{g2_timing_keyframe}()}
}
\concept{composition}
