% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_funnel}
\alias{g2_funnel}
\title{g2_funnel}
\usage{
g2_funnel(
  data,
  x,
  y,
  series = NULL,
  isTransposed = FALSE,
  shape = "funnel",
  maxSize = 1,
  minSize = 0,
  showFacetTitle = TRUE
)
}
\arguments{
\item{isTransposed}{Whether the plot is transposed.}

\item{shape}{'funnel' ｜ 'pyramid'}

\item{maxSize}{the max size of graphic，is between 0 and 1, default 1}

\item{minSize}{the min size of graphic，is between 0 and 1, default 0}

\item{showFacetTitle}{}
}
\description{
g2_funnel
}
