% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2_legend.R
\name{g2_legend}
\alias{g2_legend}
\title{g2_legend}
\usage{
g2_legend(opt, legend)
}
\arguments{
\item{opt}{G2Plot options}

\item{legend}{FALSE to close legend, or LegendCfg}
}
\value{
opt
}
\description{
g2_legend
}
\examples{
opt = list(percent=0.75)
g2_legend(opt, FALSE)
}
