% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{g2_opt}
\alias{g2_opt}
\title{g2_opt}
\usage{
g2_opt(opt, ...)
}
\arguments{
\item{opt}{G2Plot options}

\item{...}{more options}
}
\description{
g2_opt
}
\examples{
g2_opt(legend=FALSE)
g2_opt(legend=FALSE, label=FALSE)

g2_opt(
  color='rgb(252, 253, 191)-rgb(231, 82, 99)-rgb(183, 55, 121)',
  pointStyle=list(
    stroke='rgb(183, 55, 121)',
    lineWidth= 0.5
  )
)

}
