% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{g2_opt_from_json}
\alias{g2_opt_from_json}
\title{g2_opt_from_json}
\usage{
g2_opt_from_json(json)
}
\arguments{
\item{json}{a json string of G2Plot options}
}
\value{
a list of options for G2Plot
}
\description{
g2_opt_from_json
}
\examples{
opt = opt_from_json('{"xField":"displ", "yField":"hwy", "colorField":"class", "shape":"circle"}')
}
