% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{g2_output}
\alias{g2_output}
\title{Wrapper functions for using rg2 in shiny}
\usage{
g2_output(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{the width and height of the chart}
}
\description{
Use \code{g2Output()} to create a UI element, and \code{renderG2()}
to render the map widget.
}
