% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_pie}
\alias{g2_pie}
\title{g2_pie}
\usage{
g2_pie(
  data,
  angle,
  color,
  radius = NULL,
  innerRadius = NULL,
  startAngle = NULL,
  endAngle = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{angle}{angleField}

\item{color}{colorField}

\item{radius}{The radius of the pie chart, the origin is the center of the canvas. The configuration range is (0,1), where 1 represents a pie chart that fills the drawing area.}

\item{innerRadius}{The inner radius of the pie chart, starting at the center of the canvas. Configure the range (0,1)}

\item{startAngle}{Configure the starting angle of the coordinate system. try \code{pi}.}

\item{endAngle}{Configure the end angle of the coordinate system.}
}
\description{
g2_pie
}
\examples{

cfg = list(
  xField='drat',
  yField='wt',
  shape='circle',
  legend = list(position = 'top-left'),
  theme='dark' # default / dark
)
g2(mtcars, 'scatter', cfg = cfg)
g2_scatter(mtcars, 'drat', 'wt', shape='circle')


}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_line}()},
\code{\link{g2}()}
}
\concept{g2}
