% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-mark.R
\name{g2_range_x}
\alias{g2_range_x}
\title{rangeX}
\usage{
g2_range_x(data, ...)
}
\arguments{
\item{data}{to draw a single range, use a vector of length-2,to draw n ranges, use a vector of length-n, where each element of the vector is a vector of length-2}

\item{...}{}
}
\description{
rangeX
}
\details{
data for single range: \code{c(start_x1, end_x1)}

data for multiple ranges: \code{list(c(start_x1, end_x1), c(start_X2, end_x2), ...)}
}
\examples{
g2_range_x(c(67, 120))
g2_range_x(c(js('new Date("2010")'),js('new Date("2020")')))
g2_range_x(list(c(9, 18), c(24, 32), c(67,100)))

g2_view(url, children = list(g2_point(encode('height', 'weight', color='gender')),g2_range_x(c(160,170))))


}
\seealso{
Other mark: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_box}()},
\code{\link{g2_boxplot}()},
\code{\link{g2_cell}()},
\code{\link{g2_chord}()},
\code{\link{g2_col}()},
\code{\link{g2_density}()},
\code{\link{g2_gauge}()},
\code{\link{g2_heatmap}()},
\code{\link{g2_hist}()},
\code{\link{g2_image}()},
\code{\link{g2_interval}()},
\code{\link{g2_line}()},
\code{\link{g2_lineX}()},
\code{\link{g2_lineY}()},
\code{\link{g2_link}()},
\code{\link{g2_liquid}()},
\code{\link{g2_pie}()},
\code{\link{g2_point}()},
\code{\link{g2_polygon}()},
\code{\link{g2_range}()},
\code{\link{g2_range_y}()},
\code{\link{g2_rect}()},
\code{\link{g2_shape}()},
\code{\link{g2_text}()},
\code{\link{g2_vector}()},
\code{\link{g2_wordCloud}()}
}
\concept{mark}
