% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_repeat_matrix}
\alias{g2_repeat_matrix}
\title{repeatMatrix}
\usage{
g2_repeat_matrix(
  data,
  encode,
  children,
  title,
  scale,
  padding = 0,
  paddingLeft = 0,
  paddingRight = 0,
  paddingTop = 0,
  paddingBottom = 0,
  margin = 0,
  marginLeft = 0,
  marginRight = 0,
  marginTop = 0,
  marginBottom = 0
)
}
\description{
repeatMatrix
}
\seealso{
Other composition: 
\code{\link{g2_facet_circle}()},
\code{\link{g2_facet_rect}()},
\code{\link{g2_space_flex}()},
\code{\link{g2_space_layer}()},
\code{\link{g2_timing_keyframe}()}
}
\concept{composition}
