% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_ring_progress}
\alias{g2_ring_progress}
\title{g2_ring_progress}
\usage{
g2_ring_progress(percent)
}
\arguments{
\item{color}{color vector of length 2}
}
\description{
g2_ring_progress
}
\examples{
g2_ring_progress(0.75) |>
g2_opt(
  color=c('#5B8FF9', '#E8EDF3'),
  radius=0.8,
  innerRadius=0.75,
  autoFit=FALSE,
  width=200,
  height=200) |>
g2()

}
