% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_sankey}
\alias{g2_sankey}
\title{g2_sankey}
\usage{
g2_sankey(data, sourceField, targetField, weightField, rawFields = NULL)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{sourceField}{source field}

\item{targetField}{target field}

\item{weightField}{weight field}

\item{rawFields}{raw fields}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt using function with \code{g2_} prefix and \code{g2_opt()}.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
\examples{
data = jsonlite::fromJSON('[
{ "source": "首次打开","target": "首页 UV", "value": 160 },
{ "source": "结果页",  "target": "首页 UV", "value": 40 },
{ "source": "验证页",  "target": "首页 UV", "value": 10 },
{ "source": "我的",    "target": "首页 UV", "value": 10 },
{ "source": "朋友",    "target": "首页 UV", "value": 8 },
{ "source": "其他",    "target": "首页 UV", "value": 27 },
{ "source": "首页 UV", "target": "理财",    "value": 30 },
{ "source": "首页 UV", "target": "扫一扫",  "value": 40 },
{ "source": "首页 UV", "target": "服务",    "value": 35 },
{ "source": "首页 UV", "target": "蚂蚁森林","value": 25 },
{ "source": "首页 UV", "target": "跳失",    "value": 10 },
{ "source": "首页 UV", "target": "借呗",    "value": 30 },
{ "source": "首页 UV", "target": "花呗",    "value": 40 },
{ "source": "首页 UV", "target": "其他流向","value": 45 }
]')
g2_sankey(data, 'source', 'target', 'value') |> g2()

g2_sankey(data, 'source', 'target', 'value') |> g2_opt(nodeWidthRatio=0.008,nodePaddingRatio=0.03) |> g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_bidirectional_bar}()},
\code{\link{g2_box}()},
\code{\link{g2_bullet}()},
\code{\link{g2_chord}()},
\code{\link{g2_circle_packing}()},
\code{\link{g2_column}()},
\code{\link{g2_dual_axes}()},
\code{\link{g2_funnel}()},
\code{\link{g2_gauge}()},
\code{\link{g2_heatmap}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_liquid}()},
\code{\link{g2_pie}()},
\code{\link{g2_progress}()},
\code{\link{g2_radar}()},
\code{\link{g2_radial_bar}()},
\code{\link{g2_ring_progress}()},
\code{\link{g2_rose}()},
\code{\link{g2_scatter}()},
\code{\link{g2_stock}()},
\code{\link{g2_sunburst}()},
\code{\link{g2_tiny_area}()},
\code{\link{g2_tiny_column}()},
\code{\link{g2_treemap}()},
\code{\link{g2_venn}()},
\code{\link{g2_violin}()},
\code{\link{g2_waterfall}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2}()}
}
\concept{g2}
