% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_scatter}
\alias{g2_scatter}
\title{g2_scatter}
\usage{
g2_scatter(
  data,
  xField,
  yField,
  sizeField = NULL,
  shapeField = NULL,
  colorField = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{xField}{the name of the data field corresponding to the graph in the x direction. you should config the \code{opt} param with \code{xField} in \code{g2()} function, and \code{x} (short for \code{xField}) is param name for functions with \code{g2_} prefix.}

\item{yField}{the name of the data field corresponding to the graph in the y direction.}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{
g2_scatter(cars, 'speed', 'dist') |> g2()
g2_scatter(iris, 'Sepal.Length', 'Sepal.Width',colorField='Species') |> g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_funnel}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2plot}()}
}
\concept{g2}
