% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_space_flex}
\alias{g2_space_flex}
\title{spaceFlex}
\usage{
g2_space_flex(data, children = NULL, ratio, direction = "row")
}
\description{
spaceFlex
}
\seealso{
Other composition: 
\code{\link{g2_facet_circle}()},
\code{\link{g2_facet_rect}()},
\code{\link{g2_repeat_matrix}()},
\code{\link{g2_space_layer}()},
\code{\link{g2_timing_keyframe}()}
}
\concept{composition}
