% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_timing_keyframe}
\alias{g2_timing_keyframe}
\title{Timing Keyframe}
\usage{
g2_timing_keyframe(
  duration = 1000,
  iterationCount = 1,
  direction = "normal",
  children = NULL
)
}
\arguments{
\item{duration}{1000}

\item{iterationCount}{1}

\item{direction}{normal, 'normal' | 'reverse' | 'alternate' | 'reverse-alternate'}

\item{children}{marks children}
}
\description{
Timing Keyframe
}
\seealso{
Other composition: 
\code{\link{g2_facet_circle}()},
\code{\link{g2_facet_rect}()},
\code{\link{g2_repeat_matrix}()},
\code{\link{g2_space_flex}()},
\code{\link{g2_space_layer}()}
}
\concept{composition}
