% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_tiny_column}
\alias{g2_tiny_column}
\title{g2_tiny_column}
\usage{
g2_tiny_column(data)
}
\description{
g2_tiny_column
}
\examples{
g2_tiny_column(rnorm(100)) |>
g2_opt(
  columnWidthRatio=0.5,
  height=80,
  color=NULL,
  autoFit=FALSE) |>
g2()

}
