% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_tiny_line}
\alias{g2_tiny_line}
\title{g2_tiny_line}
\usage{
g2_tiny_line(data)
}
\arguments{
\item{data}{number vector}
}
\description{
g2_tiny_line
}
\examples{
g2_tiny_line(rnorm(100)) |>
g2_style(
  smooth=FALSE,
  connectNulls=TRUE,
  height=80,
  color='red',
  autoFit=FALSE) |>
g2()

}
