% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-view.R
\name{g2_view}
\alias{g2_view}
\title{View}
\usage{
g2_view(
  data,
  encode = NULL,
  children = NULL,
  scale = NULL,
  transform = NULL,
  coordinate = NULL,
  style = NULL,
  labelTransform = NULL,
  title = NULL,
  axis = NULL,
  legend = NULL,
  tooltip = NULL,
  scrollbar = NULL,
  slider = NULL,
  interaction = NULL,
  theme = NULL
)
}
\arguments{
\item{theme}{}
}
\description{
View
}
