% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2area.R
\name{g2Area}
\alias{g2Area}
\title{Area chart}
\usage{
g2Area(
  data,
  xField,
  yField,
  seriesField = NULL,
  isStack = TRUE,
  isPercent = TRUE,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, seriesField}{column name in data for aesthetic mapping}

\item{isStack, isPercent}{default to TRUE}

\item{color}{Color vector or javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Area chart
}
\seealso{
Other area: 
\code{\link{g2DualAxes}()}
}
\concept{area}
