% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2bar.R
\name{g2Bar}
\alias{g2Bar}
\title{Bar chart}
\usage{
g2Bar(
  data,
  xField,
  yField,
  seriesField = NULL,
  groupField = NULL,
  isGroup = FALSE,
  isStack = FALSE,
  isRange = FALSE,
  isPercent = FALSE,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, seriesField, groupField}{column name in data for aesthetic mapping.}

\item{isGroup, isStack, isRange, isPercent}{default to FALSE}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Bar chart
}
\seealso{
Other bar: 
\code{\link{g2GroupedBar}()}
}
\concept{bar}
