% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2BidirectionalBar.R
\name{g2BidirectionalBar}
\alias{g2BidirectionalBar}
\title{BidirectionalBar chart}
\usage{
g2BidirectionalBar(
  data,
  xField,
  yField,
  layout = "horizontal",
  yAxis = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField}{column name in data for aesthetic mapping.}

\item{layout}{Layout modes, optional 'horizontal' | 'vertical' default: 'horizontal'}

\item{yAxis}{yAxis is a list, Keys are the fields defined in yField, values is the configuration of axis.}

\item{color}{color}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
BidirectionalBar chart
}
\seealso{
Other bar: 
\code{\link{g2Bar}()},
\code{\link{g2GroupedBar}()}
}
\concept{bar}
