% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2box.R
\name{g2Box}
\alias{g2Box}
\title{Box chart}
\usage{
g2Box(
  data,
  xField,
  yField,
  groupField = NULL,
  outliersField = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, groupField, outliersField}{column name in data for aesthetic mapping.}

\item{color}{color}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Box chart
}
\concept{box}
