% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Bullet.R
\name{g2Bullet}
\alias{g2Bullet}
\title{Bullet chart}
\usage{
g2Bullet(
  data,
  rangeMax,
  measureSize = 20,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{data should contain the following columns:
* title: title character of each sample
* measures: number vector, measure of one or multiple stages
* targets: number vector, target values
* ranges: (optional) number vector, color range of the progress bar, value should be a number between 0 and 1}

\item{rangeMax}{the max value of color range of the progress bar}

\item{measureSize}{the height of progress bar, default to 20}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Bullet chart
}
\examples{
\dontrun{

# basic example for one sample
df = tibble(title='satisfaction', 
            measures=83,
            targets=90
)
g2Bullet(df, rangeMax=100)

# one sample with multiple stages(bad -> good -> excellent)
df = tibble(title='satisfaction', 
            measures=83,
            targets=90,
            ranges=list(c(0, 0.6, 0.9, 1))
)
cfg = list(
  rangeColors=c('#EE484C', '#F4A53C', '#84C14C')
)
g2Bullet(df, rangeMax=100, cfg=cfg)

# one sample with multiple targets
df = tibble(title='satisfaction', 
            measures=83, 
            targets=list(c(80, 90)),
            ranges=list(c(0, 0.6, 0.9, 1))
)
g2Bullet(df, rangeMax=100)

# group bullet chart
df = tibble(title=c('python','clojure','R', 'scala'), 
            measures=c(80,70,75,82),
            targets=c(90,90,85,85)
)
g2Bullet(df, rangeMax=100)

# group bullet chart with color range
df = tibble(title=c('python','clojure','R'), 
            measures=c(80,70,75),
            targets=c(90,90,90),
            ranges = rep(list(ranges = c(0,0.25,0.5,1)),3)
)
g2Bullet(df, rangeMax=100, cfg=list(rangeColors=c('#EE484C', '#F4A53C', '#84C14C')))

}

}
