% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2DensityHeatmap.R
\name{g2DensityHeatmap}
\alias{g2DensityHeatmap}
\title{DensityHeatmap chart}
\usage{
g2DensityHeatmap(
  data,
  xField,
  yField,
  colorField,
  color = NULL,
  radius = NULL,
  intensity = 2,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{color}{character color vector. For example: c('#295599', '#3e94c0', '#78c6d0', '#b4d9e4', '#fffef0', '#f9cdac', '#ec7d92', '#bc448c')}

\item{radius}{Numeric heat radius. automatic calculation.}

\item{intensity}{Numeric heat weight, default to 2.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
DensityHeatmap chart
}
\seealso{
Other heatmap: 
\code{\link{g2Heatmap}()}
}
\concept{heatmap}
