% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Donut.R
\name{g2Donut}
\alias{g2Donut}
\title{Donut chart}
\usage{
g2Donut(
  data,
  angleField,
  colorField = NULL,
  radius = 0.8,
  innerRadius = 0.8,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{angleField, colorField}{column name in data for aesthetic mapping}

\item{radius, innerRadius}{default to 0.8}

\item{color}{color vector or a javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Donut chart
}
\concept{}
