% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2DualAxes.R
\name{g2DualAxes}
\alias{g2DualAxes}
\title{DualAxes chart}
\usage{
g2DualAxes(
  data,
  xField,
  yField1,
  yField2,
  type = "ll",
  geometryOptions = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField1, yField2}{column name in data for aesthetic mapping}

\item{type}{ll for line-line, cl for column-line}

\item{geometryOptions}{check \url{https://g2plot.antv.vision/en/docs/api/plots/dual-axes/#geometryoptions} for more dtail.
column geometry can specify: seriesField,groupField,isStack,isGroup and some other param in g2Column, for example `list(list(geometry='column', isStack=TRUE), list(geometry='line'))`
line geometry can specify: seriesField,smooth and some other param in g2Line, for example `list(list(geometry='column'), list(geometry='line',seriesField='cate'))`}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
DualAxes chart
}
\seealso{
Other area: 
\code{\link{g2Area}()}
}
\concept{area}
