% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2DualLine.R
\name{g2DualLine}
\alias{g2DualLine}
\title{DualLine chart}
\usage{
g2DualLine(
  data,
  xField,
  yField1,
  yField2,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField1, yField2}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
DualLine chart
}
