% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2FanGauge.R
\name{g2FanGauge}
\alias{g2FanGauge}
\title{FanGauge chart}
\usage{
g2FanGauge(
  value,
  range,
  min = 0,
  max = 1,
  color = NULL,
  rangeSize = 70,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{value}{Numeric value}

\item{range}{color range of the gauge, it separate the gauge to several parts by color, example: \code{range=c(0,25,50,75,100)}}

\item{min, max}{minimum and maximum value of the chart}

\item{color}{color vector, support gradient color: `c('l(0) 0:#5d7cef 1:#e35767')`}

\item{rangeSize}{The width of gauge}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
FanGauge chart
}
\seealso{
Other gauge: 
\code{\link{g2Gauge}()},
\code{\link{g2MeterGauge}()}
}
\concept{gauge}
