% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2GroupedColumn.R
\name{g2GroupedColumn}
\alias{g2GroupedColumn}
\title{GroupedColumn chart}
\usage{
g2GroupedColumn(
  data,
  xField,
  yField,
  groupField,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, groupField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
GroupedColumn chart
}
\seealso{
Other column: 
\code{\link{g2Column}()},
\code{\link{g2PercentStackedColumn}()},
\code{\link{g2RangeColumn}()},
\code{\link{g2StackedColumn}()}
}
\concept{column}
