% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2GroupedColumnLine.R
\name{g2GroupedColumnLine}
\alias{g2GroupedColumnLine}
\title{GroupedColumnLine chart}
\usage{
g2GroupedColumnLine(
  data,
  xField,
  yField1,
  yField2,
  columnGroupField,
  lineSeriesField = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField1, yField2, columnGroupField, lineSeriesField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
GroupedColumnLine chart
}
