% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2histogram.R
\name{g2Histogram}
\alias{g2Histogram}
\title{Histogram chart}
\usage{
g2Histogram(
  data,
  binField,
  binWidth,
  binNumber = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{binField}{column name in data for aesthetic mapping}

\item{binWidth, binNumber}{width/number of bins. you must specify at least one of these.}

\item{color}{color of histogram chart}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Histogram chart
}
