% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2line.R
\name{g2Line}
\alias{g2Line}
\title{Line chart}
\usage{
g2Line(
  data,
  xField,
  yField,
  seriesField = NULL,
  smooth = FALSE,
  stepType = NULL,
  connectNulls = TRUE,
  isStack = FALSE,
  point = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, seriesField}{column name in data for aesthetic mapping}

\item{smooth}{smooth the line, default to `FALSE`}

\item{stepType}{optional hv | vh | hvh | vhv, smooth will be disabled.}

\item{connectNulls}{default to TRUE}

\item{isStack}{default to FALSE}

\item{point}{`point=list()` to display data point, Or you can specify color,size,shape and style: \url{https://g2plot.antv.vision/en/examples/line/basic/}}

\item{color}{color vector or javascript function.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Line chart
}
\concept{line}
