% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Liquid.R
\name{g2Liquid}
\alias{g2Liquid}
\title{Liquid chart}
\usage{
g2Liquid(
  value,
  min,
  max,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{value, min, max}{The current value / minimum value / maximum value}

\item{color}{The color of liquid chart}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Liquid chart
}
\concept{}
