% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2MeterGauge.R
\name{g2MeterGauge}
\alias{g2MeterGauge}
\title{MeterGauge chart}
\usage{
g2MeterGauge(
  value,
  range,
  min = 0,
  max = 1,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{value}{Numeric value}

\item{range}{color range of the gauge, it separate the gauge to several parts by color, example: \code{range=c(0,25,50,75,100)}}

\item{min, max}{minimum and maximum value of the chart}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
MeterGauge chart
}
\seealso{
Other gauge: 
\code{\link{g2FanGauge}()},
\code{\link{g2Gauge}()}
}
\concept{gauge}
