% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2PercentStackedArea.R
\name{g2PercentStackedArea}
\alias{g2PercentStackedArea}
\title{PercentStackedArea chart}
\usage{
g2PercentStackedArea(
  data,
  xField,
  yField,
  stackField,
  color = NULL,
  smooth = FALSE,
  label = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, stackField}{column name in data for aesthetic mapping}

\item{color}{Color vector or javascript function}

\item{smooth}{smooth the line}

\item{label}{The label type can be specified with `'area'` or `'line'`, default to NULL}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
PercentStackedArea chart
}
\seealso{
Other area: 
\code{\link{g2Area}()},
\code{\link{g2StackedArea}()}
}
\concept{area}
