% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2PercentStackedColumn.R
\name{g2PercentStackedColumn}
\alias{g2PercentStackedColumn}
\title{PercentStackedColumn chart}
\usage{
g2PercentStackedColumn(
  data,
  xField,
  yField,
  colorField = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{the width and height of the chart}

\item{height}{the width and height of the chart}
}
\description{
PercentStackedColumn chart
}
\seealso{
Other column: 
\code{\link{g2Column}()},
\code{\link{g2GroupedColumn}()},
\code{\link{g2RangeColumn}()},
\code{\link{g2StackedColumnLabel}()},
\code{\link{g2StackedColumn}()}
}
\concept{column}
