% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2pie.R
\name{g2Pie}
\alias{g2Pie}
\title{Pie chart}
\usage{
g2Pie(
  data,
  angleField,
  colorField = NULL,
  radius = 0.8,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{angleField, colorField}{column name in data for aesthetic mapping}

\item{radius}{default to 0.8}

\item{color}{color vector or a javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Pie chart
}
\concept{pie}
