% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Progress.R
\name{g2Progress}
\alias{g2Progress}
\title{Progress chart}
\usage{
g2Progress(
  percent,
  color = NULL,
  size = 10,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{percent}{The percent of progress, value should between 0 and 1.}

\item{color}{The color of chart}

\item{size}{The height of progress bar, default to 10.}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Progress chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2RingProgress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyColumn}()},
\code{\link{g2TinyLine}()}
}
\concept{tiny-chart}
