% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2radar.R
\name{g2Radar}
\alias{g2Radar}
\title{Radar chart}
\usage{
g2Radar(
  data,
  angleField,
  radiusField,
  seriesField = NULL,
  color = NULL,
  smooth = FALSE,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{angleField, radiusField, seriesField}{column name in data for aesthetic mapping}

\item{color}{color vector or a javascript function}

\item{smooth}{smooth the line}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Radar chart
}
