% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2RadialBar.R
\name{g2RadialBar}
\alias{g2RadialBar}
\title{RadialBar chart}
\usage{
g2RadialBar(
  data,
  xField,
  yField,
  colorField = NULL,
  type = NULL,
  radius = 1,
  innerRadius = 0.2,
  startAngle = -pi/2,
  endAngle = pi/2 * 3,
  maxAngle = 240,
  minBarWidth = NULL,
  maxBarWidth = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{type}{Display type of plot. You can specify type='line' to display a Radial-Line plot.}

\item{radius, innerRadius}{radius/innerRadius of Polar coordinate. Value can be: (0, 1]}

\item{startAngle, endAngle, maxAngle}{Angle of the coordinate system.}

\item{minBarWidth, maxBarWidth}{Specify the min/max width of bar, pixel value. Auto adapt by default.}

\item{color}{color can be a character, list or javascript function:
* '#60d7a7'
* list(rising='#60d7a7', falling='#dedede', total='#fedca9')}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
RadialBar chart
}
