% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2RangeBar.R
\name{g2RangeBar}
\alias{g2RangeBar}
\title{RangeBar chart}
\usage{
g2RangeBar(
  data,
  xField,
  yField,
  colorField = NULL,
  color = NULL,
  barSize = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping
The data in xField column should be a vector with a start and end value, \code{c(10,50)}}

\item{color}{color vector or javascript function}

\item{barSize}{automatic calculation if not specified}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
RangeBar chart
}
\seealso{
Other bar: 
\code{\link{g2Bar}()},
\code{\link{g2GroupedBar}()},
\code{\link{g2PercentStackedBar}()},
\code{\link{g2StackedBar}()}
}
\concept{bar}
