% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2RangeColumn.R
\name{g2RangeColumn}
\alias{g2RangeColumn}
\title{RangeColumn chart}
\usage{
g2RangeColumn(
  data,
  xField,
  yField,
  colorField = NULL,
  color = NULL,
  columnSize = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping
The data in yField column should be a vector with a start and end value, \code{c(10,50)}}

\item{color}{color vector or javascript function}

\item{columnSize}{automatic calculation if not specified}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
RangeColumn chart
}
\seealso{
Other column: 
\code{\link{g2Column}()},
\code{\link{g2GroupedColumn}()},
\code{\link{g2PercentStackedColumn}()},
\code{\link{g2StackedColumn}()}
}
\concept{column}
