% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2RingProgress.R
\name{g2RingProgress}
\alias{g2RingProgress}
\title{RingProgress chart}
\usage{
g2RingProgress(
  percent,
  color = NULL,
  radius = 0.98,
  innerRadius = 0.8,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{percent}{The percent of RingProgress, value should between 0 and 1.}

\item{color}{The color of chart}

\item{radius, innerRadius}{the radius/innerRadius for the RingProgress, range [0, 1]}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
RingProgress chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2Progress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyColumn}()},
\code{\link{g2TinyLine}()}
}
\concept{tiny-chart}
