% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2sankey.R
\name{g2Sankey}
\alias{g2Sankey}
\title{Sankey chart}
\usage{
g2Sankey(
  data,
  sourceField,
  targetField,
  weightField,
  nodeWidthRatio = 8e-04,
  nodeWidthPadding = 0.001,
  nodeAlign = "the justify",
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{sourceField, targetField, weightField}{column name in data for aesthetic mapping.}

\item{nodeWidthRatio}{Sankey diagram node width configuration, 0 ~ 1, refer to the width of the canvas, the default is' 0.008 '.}

\item{nodeWidthPadding}{The vertical spacing between nodes in Sankey diagram, 0 ~ 1, referring to the height of the canvas, default is' 0.01 '.}

\item{nodeAlign}{The sankey diagram node layout direction, the default is the justify, can choose the 'left' | 'right' | 'center' | 'the justify' four ways.}

\item{color}{color}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Sankey chart
}
\seealso{
Other g2: 
\code{\link{g2MultiView}()}
}
\concept{g2}
