% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2scatter.R
\name{g2Scatter}
\alias{g2Scatter}
\title{Scatter chart}
\usage{
g2Scatter(
  data,
  xField,
  yField,
  colorField = NULL,
  sizeField = NULL,
  shapeField = NULL,
  color = NULL,
  size = NULL,
  shape = "circle",
  regression = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField, sizeField, shapeField}{column name in data for aesthetic mapping}

\item{color}{points color, overwritten by colorField parameter. If there is no colorField configured, 
set one single color. Otherwise you can set a series of colors, or you can use callback function.
Default: The color board of the theme.}

\item{shape}{default to 'circle', Optional values: 'circle','square','bowtie','diamond',
'hexagon','triangle','triangle-down','cross','tick','plus','hyphen','line',
'hollow-circle','hollow-square','hollow-bowtie','hollow-diamond','hollow-hexagon',
'hollow-triangle','hollow-triangle-down'}

\item{regression}{type of regression line, exp | linear | loess | log | poly | pow | quad}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Scatter chart
}
