% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2StackedColumn.R
\name{g2StackedColumn}
\alias{g2StackedColumn}
\title{StackedColumn chart}
\usage{
g2StackedColumn(
  data,
  xField,
  yField,
  stackField,
  color = NULL,
  columnSize = NULL,
  connectedArea = FALSE,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, stackField}{column name in data for aesthetic mapping}

\item{color}{color vector or javascript function}

\item{columnSize}{automatic calculation if not specified}

\item{connectedArea}{`TRUE` fro display the connected line, default to `FALSE`}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
StackedColumn chart
}
\seealso{
Other column: 
\code{\link{g2Column}()},
\code{\link{g2GroupedColumn}()},
\code{\link{g2PercentStackedColumn}()},
\code{\link{g2RangeColumn}()}
}
\concept{column}
