% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2StackedRose.R
\name{g2StackedRose}
\alias{g2StackedRose}
\title{StackedRose chart}
\usage{
g2StackedRose(
  data,
  radiusField,
  categoryField,
  stackField,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{radiusField, categoryField, stackField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
StackedRose chart
}
\seealso{
Other rose: 
\code{\link{g2GroupedRose}()},
\code{\link{g2Rose}()}
}
\concept{rose}
