% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2StepLine.R
\name{g2StepLine}
\alias{g2StepLine}
\title{StepLine chart}
\usage{
g2StepLine(
  data,
  xField,
  yField,
  seriesField = NULL,
  step = "hv",
  color = NULL,
  lineSize = 2,
  interaction = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, seriesField}{column name in data for aesthetic mapping}

\item{step}{The step turning pattern, one of 'hv' | 'vh' | 'vhv' | 'hvh', default to 'hv'}

\item{color}{color vector or javascript function}

\item{lineSize}{default to 2}

\item{interaction}{You can specify `'slider'` or `'scrollbar'` for a large dataset}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
StepLine chart
}
\seealso{
Other line: 
\code{\link{g2Line}()}
}
\concept{line}
