% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2stock.R
\name{g2Stock}
\alias{g2Stock}
\title{Stock chart}
\usage{
g2Stock(
  data,
  xField,
  yField,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField}{column name in data for aesthetic mapping. yField should be a length-four vector for ['open', 'close', 'high', 'low']}

\item{color}{color}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Stock chart
}
\concept{stock}
