% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2sunburst.R
\name{g2Sunburst}
\alias{g2Sunburst}
\title{Sunburst chart}
\usage{
g2Sunburst(
  data,
  seriesField = NULL,
  colorField = NULL,
  type = "partition",
  radius = 1,
  innerRadius = 0.1,
  reflect = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{seriesField, colorField}{column name in data for aesthetic mapping}

\item{type}{type}

\item{radius, innerRadius}{range (0, 1]}

\item{reflect}{reflect}

\item{color}{Configure the color. If there is no colorField configured, 
set one single color. Otherwise you can set a series of colors, or you can use callback function.
Default: The color board of the theme.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Sunburst chart
}
\seealso{
Other rose: 
\code{\link{g2GroupedRose}()},
\code{\link{g2Rose}()},
\code{\link{g2StackedRose}()}
}
\concept{rose}
