% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2TinyArea.R
\name{g2TinyArea}
\alias{g2TinyArea}
\title{TinyArea chart}
\usage{
g2TinyArea(
  data,
  xField,
  yField,
  smooth = FALSE,
  color = NULL,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField}{column name in data for aesthetic mapping}

\item{smooth}{smooth the line, default to `FALSE`}

\item{color}{Color or javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
TinyArea chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2Progress}()},
\code{\link{g2RingProgress}()},
\code{\link{g2TinyColumn}()},
\code{\link{g2TinyLine}()}
}
\concept{tiny-chart}
