% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2TinyColumn.R
\name{g2TinyColumn}
\alias{g2TinyColumn}
\title{TinyColumn chart}
\usage{
g2TinyColumn(
  data,
  xField,
  yField,
  colorField = NULL,
  color = NULL,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{color}{color character, color vector or javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
TinyColumn chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2Progress}()},
\code{\link{g2RingProgress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyLine}()}
}
\concept{tiny-chart}
