% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2TinyLine.R
\name{g2TinyLine}
\alias{g2TinyLine}
\title{TinyLine chart}
\usage{
g2TinyLine(
  data,
  xField,
  yField,
  size = 2,
  smooth = FALSE,
  color = NULL,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField}{column name in data for aesthetic mapping}

\item{size}{The width of line}

\item{smooth}{smooth the line, default to `FALSE`}

\item{color}{color or javascript function.}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
TinyLine chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2Progress}()},
\code{\link{g2RingProgress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyColumn}()}
}
\concept{tiny-chart}
