% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2WordCloud.R
\name{g2WordCloud}
\alias{g2WordCloud}
\title{WordCloud chart}
\usage{
g2WordCloud(
  data,
  wordField,
  weightField,
  colorField = NULL,
  random = NULL,
  spiral = "archimedean",
  imageMask = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{wordField, weightField, colorField}{column name in data for aesthetic mapping}

\item{random}{number range (0, 1]}

\item{spiral}{optional 'archimedean' | 'rectangular' default: 'archimedean'}

\item{imageMask}{The mask image, black-white pixel image will be better, can be specified with a `url` or `base64`.}

\item{color}{Configure the color. If there is no colorField configured, 
set one single color. Otherwise you can set a series of colors, or you can use callback function.
Default: The color board of the theme.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
WordCloud chart
}
