% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-label.R
\name{label}
\alias{label}
\title{Label}
\usage{
label(...)
}
\arguments{
\item{...}{label options}
}
\description{
label is just added to the first mark
you should pass \code{labels} param to g2_ family functions if you need to define label for every mark
https://g2.antv.antgroup.com/en/spec/label/overview
}
