% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{+.g2}
\alias{+.g2}
\title{Add components to a plot}
\usage{
\method{+}{g2}(opt1, opt2)
}
\arguments{
\item{opt1}{result from \code{g2()} or  \code{+.g2}}

\item{opt2}{opt generated by \code{g2_} family (g2mark) or other spec api (g2style, ...)}
}
\description{
G2 options is named
}
\details{
\enumerate{
\item if opt2.class is g2mark then append opt into opt1.children
\item if opt1.type is spaceLayer or spaceFlex: check if opt2.encode exist
\item if opt2.class is g2spec then modify opt1 with opt2 (spec in view or composite level)
\item spec of mark will not use + operation, they configured in g2_xxx()
\item the other option of view and composite are configured by \code{g2()}
\item remove elements with null value and remove empty list
}
}
