% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderG2}
\alias{renderG2}
\title{use expr description from htmlwidgets to avoid bad inherit params code}
\usage{
renderG2(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates g2 widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
use expr description from htmlwidgets to avoid bad inherit params code
}
