% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2-package.R
\docType{package}
\name{rg2-package}
\alias{rg2}
\alias{rg2-package}
\title{rg2: G2 binding for R}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
