% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-scale.R
\name{scale}
\alias{scale}
\title{Scale}
\usage{
scale(x = NULL, y = NULL, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{z}{z}
}
\description{
Scale
}
\details{
Additional details...
\verb{[\{x: \{...\}\}, ...]}
}
