% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{spec}
\alias{spec}
\title{Spec}
\description{
G2 designed a spec to describe the visualizations that can be drawn, which
allows users to render charts by calling \code{chart.options(options)} with
specified options that meet the specification.
}
\details{
\subsection{Mark}{

\code{g2_} family make marks
}

\subsection{Transform}{

The Mark Transform in G2 offers a convenient method for transforming data and mark options, mainly used for analyzing data. The core of mark transform, which filters, modifies, aggregates and generates new channel values.

Transform is an array, executed in the order when they are declared. It can be configured in mark:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{g2_col(transform=list(list(type='stackY'), list(type='sortX')))
g2_col(transform=list(list(type='stackY')))
}\if{html}{\out{</div>}}
}
}
