% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-component.R
\name{title}
\alias{title}
\title{Title}
\usage{
title(title, subtitle = NULL, ...)
}
\description{
titles can be configured either in mark or in view
}
\seealso{
Other component: 
\code{\link{axis}()},
\code{\link{legend}()},
\code{\link{style}()}
}
\concept{component}
