% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{cfg}
\alias{cfg}
\title{G2Plot configuration
\url{https://g2plot.antv.antgroup.com/en/api}}
\description{
G2Plot configuration
\url{https://g2plot.antv.antgroup.com/en/api}
}
\examples{
\dontrun{
cfg = list(
  xField='drat',
  yField='wt',
  shape='circle',
  legend = list(position = 'top-left')
)
# theme: default / dark
cfg = list(
  theme='dark'
)
g2(mtcars, 'scatter', cfg = cfg)

}
}
