% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2.R
\name{cfg}
\alias{cfg}
\title{base configuration class}
\description{
base configuration class
}
\section{width and height}{

Both width and height are default to 400
}

\section{forceFit}{

if `TRUE`, the default, width and height in cfg will be ignored
}

\section{pixelRatio}{
 
default to 2
}

\section{renderer}{

One of the following:
\itemize{
   \item "canvas"
   \item "svg"
}
}

\section{theme}{

chart theme. One of the following:
\itemize{
  \item "default" (the default)
  \item "dark"
}
}

\section{title and description}{

Both title and description in cfg should be list.
}

\examples{
\dontrun{
# Default cfg for title and description
cfg = list(
  title = list(visible = FALSE,
               alignTo = 'left',
               text    = '',
               style   = list(fontSize=18, fill='black')
  ),
  description = lsit(visible = FALSE,
                     alignTo = 'left',
                     text    = '',
                     style   = list(fontSize=12, fill='grey')
  )
)
g2bar(df, 'x', 'y', cfg=cfg)


}
}
