% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2.R
\name{g2}
\alias{g2}
\title{base class for G2 chart}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
base class for G2 chart
}
