% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2}
\alias{g2}
\title{g2}
\usage{
g2(data, encode = NULL, ..., width = 640, height = 480)
}
\arguments{
\item{data}{required. Data Frame, URL of json/csv}

\item{encode}{encode}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{width}{optional default: 640}

\item{height}{optional default: 480}

\item{type}{view/spaceLayer/spaceFlex/facetCircle/facetRect/repeatMatrix}
}
\description{
g2() create an G2 options
}
\details{
Spec: \url{https://g2.antv.antgroup.com/en/spec/overview}
}
\examples{
g2_point(mtcars, 'drat', 'wt')

}
\concept{g2}
