% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2.R
\name{g2}
\alias{g2}
\title{base class for G2 chart}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{the width of the chart}

\item{height}{the width of the chart}
}
\description{
base class for G2 chart
}
