% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_bar}
\alias{g2_bar}
\title{g2_bar}
\usage{
g2_bar(
  data,
  x,
  y,
  series = NULL,
  group = NULL,
  color = NULL,
  isGroup = NULL,
  isStack = NULL,
  isRange = NULL,
  isPercent = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{x}{short for xField}

\item{y}{short for yField}

\item{isGroup}{to group charts}

\item{isStack}{to stack charts}

\item{isRange}{to range charts}

\item{isPercent}{to percent charts}

\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}
}
\description{
g2_bar
}
\examples{
df = jsonlite::fromJSON('
[{ "year": "1951", "value": 38 },
  { "year": "1952", "value": 52 },
  { "year": "1956", "value": 61 },
  { "year": "1957", "value": 145 },
  { "year": "1958", "value": 48 }]')
g2_bar(df, 'value', 'year')

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_column}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2_scatter}()},
\code{\link{g2}()}
}
\concept{g2}
