% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-mark.R
\name{g2_box}
\alias{g2_box}
\title{box}
\usage{
g2_box(encode = NULL, ...)
}
\arguments{
\item{encode}{encode}

\item{...}{...}
}
\description{
box is an Atomic mark, you need to calculate the quantile.
}
\details{
Quantile data for box mark:
\link{min, Q1, Median, Q3, max}
}
\seealso{
Other mark: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_boxplot}()},
\code{\link{g2_cell}()},
\code{\link{g2_chord}()},
\code{\link{g2_col}()},
\code{\link{g2_density}()},
\code{\link{g2_gauge}()},
\code{\link{g2_heatmap}()},
\code{\link{g2_hist}()},
\code{\link{g2_image}()},
\code{\link{g2_interval}()},
\code{\link{g2_line}()},
\code{\link{g2_lineX}()},
\code{\link{g2_lineY}()},
\code{\link{g2_link}()},
\code{\link{g2_liquid}()},
\code{\link{g2_pie}()},
\code{\link{g2_point}()},
\code{\link{g2_polygon}()},
\code{\link{g2_range}()},
\code{\link{g2_range_x}()},
\code{\link{g2_range_y}()},
\code{\link{g2_rect}()},
\code{\link{g2_shape}()},
\code{\link{g2_text}()},
\code{\link{g2_vector}()},
\code{\link{g2_wordCloud}()}
}
\concept{mark}
