% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_bullet}
\alias{g2_bullet}
\title{g2_bullet}
\usage{
g2_bullet(data, measureField, rangeField, targetField, xField = NULL)
}
\arguments{
\item{data}{The data source is a collection of objects [{title: 'income', ranges: \link{50,100}, measures: \link{80}, target: 85}]}

\item{measureField}{Use the length of the data bar, the setting field for the actual value, to represent the actual value.}

\item{rangeField}{Use the setting field for the length of the background bar to represent the range.}

\item{targetField}{Use a setting field for the position of the scale axis of the measurement mark to represent the target value.}

\item{xField}{Used to distinguish different types, suitable for grouping bullet diagrams.}
}
\description{
g2_bullet
}
\examples{
data = jsonlite::fromJSON('[{"title":"the title","ranges":[100],"measures":[80],"target":85}]')
g2_bullet(data,'measures','ranges','target') |>
g2_opt(color=list(range='#f0efff',measure= '#5B8FF9',target='#3D76DD'),
       size=list(range=12,measure=12,target=12)) |>
g2()

}
