% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_column}
\alias{g2_column}
\title{g2_column}
\usage{
g2_column(
  data,
  x,
  y,
  series = NULL,
  group = NULL,
  color = NULL,
  isGroup = NULL,
  isStack = NULL,
  isRange = NULL,
  isPercent = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{x}{short for xField}

\item{y}{short for yField}

\item{isGroup}{to group charts}

\item{isStack}{to stack charts}

\item{isRange}{to range charts}

\item{isPercent}{to percent charts}

\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}
}
\description{
g2_column
}
\examples{
df = jsonlite::fromJSON('https://gw.alipayobjects.com/os/antfincdn/8elHX\%26irfq/stack-column-data.json',simplifyVector = TRUE)
g2_column(df,'year','value','type',isStack=TRUE)

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2_scatter}()},
\code{\link{g2}()}
}
\concept{g2}
