% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_histogram}
\alias{g2_histogram}
\title{g2_histogram}
\usage{
g2_histogram(data, binField, stackField = NULL)
}
\arguments{
\item{data}{required. Data frame for plot.}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{
data.frame(value=rnorm(1000)) |>
g2_histogram('value') |>
g2()

data.frame(value=rnorm(1000)) |>
g2_histogram('value') |>
g2_style(binNumber=100) |>
g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_funnel}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2_scatter}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2plot}()}
}
\concept{g2}
