% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_line}
\alias{g2_line}
\title{g2_line}
\usage{
g2_line(
  data,
  x,
  y,
  series = NULL,
  color = NULL,
  smooth = FALSE,
  isStack = FALSE,
  stepType = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{x}{short for xField}

\item{y}{short for yField}

\item{smooth}{TRUE/FALSE. whether the curve is smooth.}

\item{isStack}{to stack charts}

\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}
}
\description{
g2_line
}
\examples{

df = jsonlite::read_json("https://gw.alipayobjects.com/os/bmw-prod/1d565782-dde4-4bb6-8946-ea6a38ccf184.json",simplifyVector = T)
g2_line(df, 'Date','scales')

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2_scatter}()},
\code{\link{g2}()}
}
\concept{g2}
