% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_pie}
\alias{g2_pie}
\title{g2_pie}
\usage{
g2_pie(data, angleField, colorField)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{colorField}{the name of the data field or color name or hex color code}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{
data.frame(
  cate = c('a','b','c','d','e'),
  value = c(1,2,3,5,8)
) |>
  g2_pie('value','cate') |>
  g2_opt(radius=.8, innerRadius=.4) |>
  g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_funnel}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_radar}()},
\code{\link{g2_scatter}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2plot}()}
}
\concept{g2}
