% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_progress}
\alias{g2_progress}
\title{g2_progress}
\usage{
g2_progress(percent)
}
\arguments{
\item{color}{color vector of length 2}

\item{barWidthRatio}{\link{0-1}}
}
\description{
g2_progress
}
\examples{
g2_progress(0.75) |>
g2_opt(
  color=c('#F4664A', '#E8EDF3'),
  barWidthRatio=0.3,
  autoFit=FALSE,
  width=300,
  height=100) |>
g2()

}
