% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_radar}
\alias{g2_radar}
\title{g2_radar}
\usage{
g2_radar(data, xField, yField, seriesField = NULL)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{xField}{The name of the data field corresponding to the graph in the x direction, usually the field corresponding to the horizontal coordinate axis. For example, to see how many people are in different classes, the class field is the corresponding xField.}

\item{yField}{The name of the data field corresponding to the graph in the y direction, usually the field corresponding to the vertical coordinate axis. For example, to see the number of students in different classes, the number field is the corresponding yField.}

\item{seriesField}{grouping field}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{
data = jsonlite::fromJSON('https://gw.alipayobjects.com/os/antfincdn/svFjSfJkYy/radar.json')
g2_radar(data,'item','score','user') |> g2_opt(area=list()) |> g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_funnel}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_scatter}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2plot}()}
}
\concept{g2}
