% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_repeat_matrix}
\alias{g2_repeat_matrix}
\title{repeatMatrix}
\usage{
g2_repeat_matrix(
  data,
  encode,
  children,
  title = NULL,
  scale = NULL,
  padding = 0,
  paddingLeft = 0,
  paddingRight = 0,
  paddingTop = 0,
  paddingBottom = 0,
  margin = 0,
  marginLeft = 0,
  marginRight = 0,
  marginTop = 0,
  marginBottom = 0,
  width = 640,
  height = 480
)
}
\arguments{
\item{data}{data}

\item{encode}{encode}

\item{children}{children}

\item{title}{title}

\item{scale}{scale}

\item{padding}{padding}

\item{paddingLeft}{paddingLeft}

\item{paddingRight}{paddingRight}

\item{paddingTop}{paddingTop}

\item{paddingBottom}{paddingBottom}

\item{margin}{margin}

\item{marginLeft}{marginLeft}

\item{marginRight}{marginRight}

\item{marginTop}{marginTop}

\item{marginBottom}{marginBottom}
}
\description{
repeatMatrix
}
\seealso{
Other composition: 
\code{\link{g2_facet_circle}()},
\code{\link{g2_facet_rect}()},
\code{\link{g2_space_flex}()},
\code{\link{g2_space_layer}()},
\code{\link{g2_timing_keyframe}()}
}
\concept{composition}
