% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_scatter}
\alias{g2_scatter}
\title{g2_scatter}
\usage{
g2_scatter(
  data,
  x,
  y,
  size = NULL,
  shape = NULL,
  color = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{x}{short for xField}

\item{y}{short for yField}

\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{
g2_scatter(cars, 'speed', 'dist')
g2_scatter(iris, 'Sepal.Length', 'Sepal.Width',color='Species')

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_column}()},
\code{\link{g2_line}()},
\code{\link{g2_pie}()},
\code{\link{g2_radar}()},
\code{\link{g2}()}
}
\concept{g2}
