% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{g2_scatter}
\alias{g2_scatter}
\title{g2_scatter}
\usage{
g2_scatter(
  data,
  xField,
  yField,
  sizeField = NULL,
  shapeField = NULL,
  colorField = NULL
)
}
\arguments{
\item{data}{required. Data frame for plot.}

\item{xField}{The name of the data field corresponding to the graph in the x direction, usually the field corresponding to the horizontal coordinate axis. For example, to see how many people are in different classes, the class field is the corresponding xField.}

\item{yField}{The name of the data field corresponding to the graph in the y direction, usually the field corresponding to the vertical coordinate axis. For example, to see the number of students in different classes, the number field is the corresponding yField.}

\item{sizeField}{size field}

\item{shapeField}{shape field}

\item{colorField}{color field}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt using function with \code{g2_} prefix and \code{g2_opt()}.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
\examples{
g2_scatter(cars, 'speed', 'dist') |> g2()
g2_scatter(iris, 'Sepal.Length', 'Sepal.Width',colorField='Species') |> g2()

}
\seealso{
Other g2: 
\code{\link{g2_area}()},
\code{\link{g2_bar}()},
\code{\link{g2_bidirectional_bar}()},
\code{\link{g2_box}()},
\code{\link{g2_bullet}()},
\code{\link{g2_chord}()},
\code{\link{g2_circle_packing}()},
\code{\link{g2_column}()},
\code{\link{g2_dual_axes}()},
\code{\link{g2_funnel}()},
\code{\link{g2_gauge}()},
\code{\link{g2_heatmap}()},
\code{\link{g2_histogram}()},
\code{\link{g2_line}()},
\code{\link{g2_liquid}()},
\code{\link{g2_pie}()},
\code{\link{g2_progress}()},
\code{\link{g2_radar}()},
\code{\link{g2_radial_bar}()},
\code{\link{g2_ring_progress}()},
\code{\link{g2_rose}()},
\code{\link{g2_sankey}()},
\code{\link{g2_stock}()},
\code{\link{g2_sunburst}()},
\code{\link{g2_tiny_area}()},
\code{\link{g2_tiny_column}()},
\code{\link{g2_treemap}()},
\code{\link{g2_venn}()},
\code{\link{g2_violin}()},
\code{\link{g2_waterfall}()},
\code{\link{g2_word_cloud}()},
\code{\link{g2}()}
}
\concept{g2}
