% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-composition.R
\name{g2_space_flex}
\alias{g2_space_flex}
\title{spaceFlex}
\usage{
g2_space_flex(
  data = NULL,
  children = NULL,
  ratio = NULL,
  direction = "row",
  padding = 0,
  paddingLeft = 0,
  paddingRight = 0,
  paddingTop = 0,
  paddingBottom = 0,
  margin = 0,
  marginLeft = 0,
  marginRight = 0,
  marginTop = 0,
  marginBottom = 0,
  width = 640,
  height = 480
)
}
\arguments{
\item{data}{data}

\item{children}{children marks}

\item{ratio}{ratio}

\item{direction}{direction}

\item{padding}{padding}

\item{paddingLeft}{paddingLeft}

\item{paddingRight}{paddingRight}

\item{paddingTop}{paddingTop}

\item{paddingBottom}{paddingBottom}

\item{margin}{margin}

\item{marginLeft}{marginLeft}

\item{marginRight}{marginRight}

\item{marginTop}{marginTop}

\item{marginBottom}{marginBottom}

\item{width}{width}

\item{height}{height}
}
\description{
spaceFlex
}
\seealso{
Other composition: 
\code{\link{g2_facet_circle}()},
\code{\link{g2_facet_rect}()},
\code{\link{g2_repeat_matrix}()},
\code{\link{g2_space_layer}()},
\code{\link{g2_timing_keyframe}()}
}
\concept{composition}
