% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2-view.R
\name{g2_view}
\alias{g2_view}
\title{View}
\usage{
g2_view(
  data,
  encode = NULL,
  children = NULL,
  scale = NULL,
  transform = NULL,
  coordinate = NULL,
  style = NULL,
  labelTransform = NULL,
  title = NULL,
  axis = NULL,
  legend = NULL,
  tooltip = NULL,
  scrollbar = NULL,
  slider = NULL,
  interaction = NULL,
  theme = NULL,
  width = 640,
  height = 480
)
}
\arguments{
\item{data}{data}

\item{encode}{encode}

\item{children}{children}

\item{scale}{scale}

\item{transform}{transform}

\item{coordinate}{coordinate}

\item{style}{style}

\item{labelTransform}{labelTransform}

\item{title}{title}

\item{axis}{axis}

\item{legend}{legend}

\item{tooltip}{tooltip}

\item{scrollbar}{scrollbar}

\item{slider}{slider}

\item{interaction}{interaction}

\item{theme}{theme}
}
\description{
View
}
