% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Area.R
\name{g2Area}
\alias{g2Area}
\title{Area chart}
\usage{
g2Area(
  data,
  xField,
  yField,
  seriesField = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, seriesField}{column name in data for aesthetic mapping}

\item{color}{Color vector or javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Area chart
}
\seealso{
Other area: 
\code{\link{g2PercentStackedArea}()},
\code{\link{g2StackedArea}()}
}
\concept{area}
