% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2column.R
\name{g2Column}
\alias{g2Column}
\title{Column chart}
\usage{
g2Column(
  data,
  xField,
  yField,
  colorField = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Column chart
}
\concept{column}
