% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2density.R
\name{g2Density}
\alias{g2Density}
\title{Density chart}
\usage{
g2Density(
  data,
  binField,
  binWidth = NULL,
  binNumber = NULL,
  kernel = "epanechnikov",
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{binField}{column name in data for aesthetic mapping}

\item{binWidth, binNumber}{width/number of bins. you must specify at least one of these.}

\item{kernel}{One of `'uniform'` or `'triangle'` or `'epanechnikov'` or `'quartic'` or `'triweight'` or `'gaussian'` or `'cosinus'`, default to `'epanechnikov'`.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Density chart
}
