% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2funnel.R
\name{g2Funnel}
\alias{g2Funnel}
\title{Funnel chart}
\usage{
g2Funnel(
  data,
  xField,
  yField,
  compareField = NULL,
  dynamicHeight = FALSE,
  transpose = FALSE,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField}{column name in data for aesthetic mapping}

\item{compareField}{Required by compare funnel. specified values will overwrite the dynamicHeight to `FALSE`.
you can configure the compare text style in `cfg` parameter with `compareText=list(...)`}

\item{dynamicHeight}{`TRUE` for mapping xField to width/angle of funnel with a fixed height, 
`FALSE` for mapping xField to height of funnel with a fixed angle.}

\item{transpose}{`TRUE` for horizontal funnel, and `FALSE` for vertical funnel}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Funnel chart
}
\section{percentage configuration}{

In the funnel chart, the conversion rate display consists of three parts: 
\enumerate{
  \item fixed line
  \item fixed text
  \item numeric text
}
 
 The style and content can be configured by `line`, `text` and `value` respectively.
 
 See example for a full configuration of percentage.
}

\examples{
 \dontrun{
 
 cfg = list(
   percentage = list(
     visible = TRUE,
     offsetX = 40,
     offsetY = 40,
     spacing = 4,
     line = list(
       visible = TRUE,
       style = list()  # lineStyle
     ),
     text = list(
       visible = TRUE,
       content = 'conversion rate'
       style = list()  # textStyle
     ),
     value = list(
       visible = TRUE,
       style = list()  # textStyle
     )
   )
 )
 
 }


}
