% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2funnel.R
\name{g2Funnel}
\alias{g2Funnel}
\title{Funnel chart}
\usage{
g2Funnel(
  data,
  xField,
  yField,
  compareField = NULL,
  seriesField = NULL,
  isTransposed = FALSE,
  dynamicHeight = FALSE,
  maxSize = 1,
  minSize = 1,
  conversionTag = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, compareField, seriesField}{column name in data for aesthetic mapping}

\item{isTransposed}{default to FALSE}

\item{dynamicHeight}{`TRUE` for mapping xField to width/angle of funnel with a fixed height, 
`FALSE` for mapping xField to height of funnel with a fixed angle.}

\item{maxSize, minSize}{these fields are invalid when dynamicHeight=true}

\item{conversionTag}{FALSE for hide the conversion info or specify a `htmlwidgets::JS` \url{https://g2plot.antv.vision/en/docs/api/plots/funnel/}}

\item{color}{color}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Funnel chart
}
