% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2GroupedBar.R
\name{g2GroupedBar}
\alias{g2GroupedBar}
\title{GroupedBar chart}
\usage{
g2GroupedBar(
  data,
  xField,
  yField,
  colorField = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{the width and height of the chart}

\item{height}{the width and height of the chart}
}
\description{
GroupedBar chart
}
\seealso{
Other bar: 
\code{\link{g2Bar}()},
\code{\link{g2PercentStackedBar}()},
\code{\link{g2RangeBar}()},
\code{\link{g2StackedBar}()}
}
\concept{bar}
