% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2heatmap.R
\name{g2Heatmap}
\alias{g2Heatmap}
\title{Heatmap chart}
\usage{
g2Heatmap(
  data,
  xField,
  yField,
  colorField = NULL,
  sizeField = NULL,
  type = "polygon",
  reflect = "x",
  color = NULL,
  shape = "rect",
  sizeRatio = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField, sizeField}{column name in data for aesthetic mapping}

\item{type}{optional polygon | density default: polygon}

\item{reflect}{Axis mapping, optional x | y}

\item{color}{color range vector}

\item{shape}{optional rect | square | circle}

\item{sizeRatio}{The scale of the shapes in the thermal lattice, optional, takes effect only if the shape and sizeField specify at least one of these.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Heatmap chart
}
\seealso{
Other heatmap: 
\code{\link{g2DensityHeatmap}()}
}
\concept{heatmap}
