% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Histogram.R
\name{g2Histogram}
\alias{g2Histogram}
\title{Histogram chart}
\usage{
g2Histogram(
  data,
  xField,
  yField,
  colorField = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{the width and height of the chart}

\item{height}{the width and height of the chart}
}
\description{
Histogram chart
}
\concept{}
