% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Histogram.R
\name{g2Histogram}
\alias{g2Histogram}
\title{Histogram chart}
\usage{
g2Histogram(
  data,
  xField,
  yField,
  colorField = "",
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{the width and height of the chart}

\item{height}{the width and height of the chart}
}
\description{
Histogram chart
}
\seealso{
Other bar: 
\code{\link{g2Area}()},
\code{\link{g2Bar}()},
\code{\link{g2Bubble}()},
\code{\link{g2Bullet}()},
\code{\link{g2Calendar}()},
\code{\link{g2ColumnLabel}()},
\code{\link{g2Column}()},
\code{\link{g2DensityHeatmap}()},
\code{\link{g2Density}()},
\code{\link{g2Donut}()},
\code{\link{g2FanGauge}()},
\code{\link{g2Funnel}()},
\code{\link{g2Gauge}()},
\code{\link{g2GroupedBar}()},
\code{\link{g2GroupedColumn}()},
\code{\link{g2GroupedRose}()},
\code{\link{g2Heatmap}()},
\code{\link{g2Line}()},
\code{\link{g2Liquid}()},
\code{\link{g2MeterGauge}()},
\code{\link{g2PercentStackedArea}()},
\code{\link{g2PercentStackedBar}()},
\code{\link{g2PercentStackedColumn}()},
\code{\link{g2Pie}()},
\code{\link{g2Radar}()},
\code{\link{g2RangeBar}()},
\code{\link{g2RangeColumn}()},
\code{\link{g2Rose}()},
\code{\link{g2Scatter}()},
\code{\link{g2StackedArea}()},
\code{\link{g2StackedBar}()},
\code{\link{g2StackedColumnLabel}()},
\code{\link{g2StackedColumn}()},
\code{\link{g2StackedRose}()},
\code{\link{g2StepLine}()},
\code{\link{g2Treemap}()},
\code{\link{g2Waterfall}()},
\code{\link{g2WordCloud}()}
}
\concept{bar}
