% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2liquid.R
\name{g2liquid}
\alias{g2liquid}
\title{Liquid chart}
\usage{
g2liquid(
  data,
  xField,
  yField,
  colorField = "",
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{colname in data for aesthetic mapping}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{the width of the chart}

\item{height}{the width of the chart}
}
\description{
Liquid chart
}
\seealso{
Other bar: 
\code{\link{g2area}()},
\code{\link{g2bar}()},
\code{\link{g2bubble}()},
\code{\link{g2bullet}()},
\code{\link{g2calendar}()},
\code{\link{g2column_label}()},
\code{\link{g2column}()},
\code{\link{g2density_heatmap}()},
\code{\link{g2density}()},
\code{\link{g2donut}()},
\code{\link{g2fan_gauge}()},
\code{\link{g2funnel}()},
\code{\link{g2gauge}()},
\code{\link{g2grouped_bar}()},
\code{\link{g2grouped_column}()},
\code{\link{g2grouped_rose}()},
\code{\link{g2heatmap}()},
\code{\link{g2histogram}()},
\code{\link{g2line}()},
\code{\link{g2meter_gauge}()},
\code{\link{g2percent_stacked_area}()},
\code{\link{g2percent_stacked_bar}()},
\code{\link{g2percent_stacked_column}()},
\code{\link{g2pie}()},
\code{\link{g2radar}()},
\code{\link{g2range_bar}()},
\code{\link{g2range_column}()},
\code{\link{g2rose}()},
\code{\link{g2scatter}()},
\code{\link{g2stacked_area}()},
\code{\link{g2stacked_bar}()},
\code{\link{g2stacked_column_label}()},
\code{\link{g2stacked_column}()},
\code{\link{g2stacked_rose}()},
\code{\link{g2step_line}()},
\code{\link{g2treemap}()},
\code{\link{g2waterfall}()},
\code{\link{g2word_cloud}()}
}
\concept{bar}
