% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2liquid.R
\name{g2Liquid}
\alias{g2Liquid}
\title{Liquid chart}
\usage{
g2Liquid(
  percent,
  radius = 0.9,
  color = NULL,
  outline = list(border = 2, distance = 0),
  wave = list(count = 3, length = 192),
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{percent}{percent value}

\item{radius}{radius [0, 1]}

\item{color}{The color of liquid chart}

\item{outline}{default `list(border=2, distance=0)`}

\item{wave}{default `list(count=3, length=192)`}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Liquid chart
}
