% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2MultiView.R
\name{g2MultiView}
\alias{g2MultiView}
\title{MultiView chart}
\usage{
g2MultiView(data, cfg = list(), width = NULL, height = NULL)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
MultiView chart
}
\seealso{
Other g2: 
\code{\link{g2Sankey}()}
}
\concept{g2}
