% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2MultiView.R
\name{g2MultiView}
\alias{g2MultiView}
\title{MultiView chart}
\usage{
g2MultiView(cfg = list(), width = NULL, height = NULL)
}
\arguments{
\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
MultiView chart
}
\examples{
\dontrun{

df = jsonlite::fromJSON('https://gw.alipayobjects.com/os/antfincdn/fKTgtjKdaN/association-pie.json')
cfg = list(
  tooltip=FALSE,
  legend=TRUE,
  plots=list(
    list(
      type='pie',
      region=list(start=list(x=0, y=0 ), end=list(x=0.45, y=1 ) ),
      options=list(
        data=jsonlite::toJSON(df$pie1),
        angleField='bill',
        colorField='area',
        radius=0.85,
        tooltip=list(showMarkers=FALSE),
        label=list(type='inner', offset='-15\%'),
        interactions=list(
          type=list('element-active'),
          type=list('association-tooltip'),
          type=list('association-highlight')
        )
      )
    ),
    list(
      type='pie',
      region=list(start=list(x=0.55, y=0 ), end=list(x=1, y=1 ) ),
      options=list(
        data=jsonlite::toJSON(df$pie2),
        angleField='value',
        colorField='area',
        radius=0.85,
        tooltip=list(showMarkers=FALSE),
        label=list(type='inner', offset='-15\%'),
        interactions=list(
          type=list('association-tooltip'),
          type=list('association-selected')
        )
      )
    )
  )
)

g2MultiView(cfg)

}

}
\seealso{
Other g2: 
\code{\link{g2Sankey}()}
}
\concept{g2}
