% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Progress.R
\name{g2Progress}
\alias{g2Progress}
\title{Progress chart}
\usage{
g2Progress(
  percent,
  barWidthRatio = 10,
  color = NULL,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{percent}{The percent of progress, value should between 0 and 1.}

\item{barWidthRatio}{Ragne [0, 1], defualt 0.5}

\item{color}{The color of chart, single color or length-two color vector}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Progress chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2RingProgress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyColumn}()},
\code{\link{g2TinyLine}()}
}
\concept{tiny-chart}
