% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2rose.R
\name{g2Rose}
\alias{g2Rose}
\title{Rose chart}
\usage{
g2Rose(
  data,
  xField,
  yField,
  seriesField = NULL,
  radius = NULL,
  innerRadius = NULL,
  color = NULL,
  isGroup = FALSE,
  isStack = FALSE,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{radius, innerRadius}{range (0, 1]}

\item{color}{Configure the color. If there is no colorField configured, 
set one single color. Otherwise you can set a series of colors, or you can use callback function.
Default: The color board of the theme.}

\item{isGroup, isStack}{default to FALSE}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Rose chart
}
\seealso{
Other rose: 
\code{\link{g2GroupedRose}()},
\code{\link{g2StackedRose}()}
}
\concept{rose}
