% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2TinyColumn.R
\name{g2TinyColumn}
\alias{g2TinyColumn}
\title{TinyColumn chart}
\usage{
g2TinyColumn(
  data,
  columnWidthRatio = 0.5,
  color = NULL,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{data}{numeric vector}

\item{columnWidthRatio}{optional number default: 0.5}

\item{color}{color character, color vector or javascript function}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api/} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
TinyColumn chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2Progress}()},
\code{\link{g2RingProgress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyLine}()}
}
\concept{tiny-chart}
