% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2TinyLine.R
\name{g2TinyLine}
\alias{g2TinyLine}
\title{TinyLine chart}
\usage{
g2TinyLine(
  data,
  smooth = FALSE,
  connectNulls = FALSE,
  color = NULL,
  cfg = list(),
  width = 200,
  height = 50
)
}
\arguments{
\item{data}{numeric vector}

\item{smooth}{smooth the line, default to `FALSE`}

\item{connectNulls}{The width of line}

\item{color}{color or javascript function.}

\item{cfg}{configuration of G2Plot. See \url{https://g2plot.antv.vision/en/docs/api/plot-api} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
TinyLine chart
}
\seealso{
Other tiny-chart: 
\code{\link{g2Progress}()},
\code{\link{g2RingProgress}()},
\code{\link{g2TinyArea}()},
\code{\link{g2TinyColumn}()}
}
\concept{tiny-chart}
