% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Treemap.R
\name{g2Treemap}
\alias{g2Treemap}
\title{Treemap chart}
\usage{
g2Treemap(data, maxLevel = 2, cfg = list(), width = NULL, height = NULL)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{maxLevel}{Default to 2}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Treemap chart
}
