% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2Waterfall.R
\name{g2Waterfall}
\alias{g2Waterfall}
\title{Waterfall chart}
\usage{
g2Waterfall(
  data,
  xField,
  yField,
  colorField = NULL,
  color = NULL,
  cfg = list(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{a data object.Currently supported object is data.frame, 
data should only contains the necessary data for chart (reduce json/data transmission)}

\item{xField, yField, colorField}{column name in data for aesthetic mapping}

\item{color}{color can be a character, list or javascript function:
* '#60d7a7'
* list(rising='#60d7a7', falling='#dedede', total='#fedca9')}

\item{cfg}{configuration of G2Plot. See \url{https://antv-g2plot.gitee.io/zh/examples/general/axis} and \url{https://antv-g2plot.gitee.io/zh/docs/manual/graphic-style} for more details.}

\item{width}{width of the chart}

\item{height}{height of the chart}
}
\description{
Waterfall chart
It's not recommended to use colorField param for waterfall chart (such as basic the column chart)
}
