% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{opt}
\alias{opt}
\title{G2 options}
\arguments{
\item{width}{optional default: 640}

\item{height}{optional default: 480}

\item{data}{required. Data Frame, URL of json/csv}
}
\description{
G2 provides a Spec API which declares visualizations through a JavaScript object.

Options of G2 is a named list:

\if{html}{\out{<div class="sourceCode">}}\preformatted{\{
  type: 'mark',  // view, spaceLayer, ...
  data: [],
  encode: \{\},
  scale: \{\},
  transform: [],
  coordinate: \{\},
  style: \{\},
  labelTransform: \{\},
  title: \{\},
  axis: \{\},
  legend: \{\},
  tooltip: \{\},
  scrollbar: \{\},
  slider: \{\},
  interaction: \{\},
  theme: \{\},
\}
}\if{html}{\out{</div>}}

For the option with array value, you should use \code{list(list(...), list(...))}
to construct an array in JSON.

\if{html}{\out{<div class="sourceCode">}}\preformatted{transform()
}\if{html}{\out{</div>}}
}
\details{
Spec: \url{https://g2.antv.antgroup.com/en/spec/overview}
}
\examples{
g2_point(mtcars, 'drat', 'wt')

}
