% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg2.R
\name{opt}
\alias{opt}
\title{G2Plot options}
\arguments{
\item{width}{optional number default: 400. width of the chart.}

\item{height}{optional number default: 400. height of the chart.}

\item{data}{required. Data frame for plot.}

\item{xField}{the name of the data field corresponding to the graph in the x direction. you should config the \code{opt} param with \code{xField} in \code{g2()} function, and \code{x} (short for \code{xField}) is param name for functions with \code{g2_} prefix.}

\item{yField}{the name of the data field corresponding to the graph in the y direction.}

\item{x}{short for xField}

\item{y}{short for yField}

\item{color, colorField}{the name of the data field or color name or hex color code}

\item{group, groupField}{grouping field}

\item{series, seriesField}{grouping field}

\item{smooth}{TRUE/FALSE. whether the curve is smooth.}

\item{isStack}{to stack charts}

\item{isGroup}{to group charts}

\item{isPercent}{to percent charts}

\item{isRange}{to range charts}
}
\description{
opt param in rg2 is the PlotOptions for G2Plot.
You can construct an opt in two way:
\itemize{
\item use \code{list()} function in R base package.
\item use \code{opt_from_json()} from rg2 package.
\url{https://g2plot.antv.antgroup.com/en/api} for details
}
}
\examples{

opt = list(
  xField='drat',
  yField='wt',
  shape='circle',
  legend = list(position = 'top-left'),
  theme='dark' # default / dark
)
g2(mtcars, 'scatter', opt = opt)

opt = opt_from_json('{"xField":"drat","yField":"wt", "shape":"circle","lengend":{"position":"top-left"},"theme":"dark"}')
g2(mtcars, 'scatter', opt = opt)

g2_scatter(mtcars, 'drat', 'wt', shape='circle')


}
